Fix some model issues

+++ OBSWebSocketLibrary/Models/TypeDefs/SceneItem.cs
@@ -8,15 +8,15 @@ namespace OBSWebSocketLibrary.Models.TypeDefs
     public class SceneItem
     {
         [JsonPropertyName("cy")]
-        public int Cy { get; set; }
+        public double Cy { get; set; }
         [JsonPropertyName("cx")]
-        public int Cx { get; set; }
+        public double Cx { get; set; }
         [JsonPropertyName("alignment")]
-        public int Alignment { get; set; }
+        public double Alignment { get; set; }
         [JsonPropertyName("name")]
         public string Name { get; set; }
         [JsonPropertyName("id")]
-        public int Id { get; set; }
+        public double Id { get; set; }
         [JsonPropertyName("render")]
         public bool Render { get; set; }
         [JsonPropertyName("muted")]
@@ -24,17 +24,17 @@ namespace OBSWebSocketLibrary.Models.TypeDefs
         [JsonPropertyName("locked")]
         public bool Locked { get; set; }
         [JsonPropertyName("source_cx")]
-        public int SourceCx { get; set; }
+        public double SourceCx { get; set; }
         [JsonPropertyName("source_cy")]
-        public int SourceCy { get; set; }
+        public double SourceCy { get; set; }
         [JsonPropertyName("type")]
         public string Type { get; set; }
         [JsonPropertyName("volume")]
-        public int Volume { get; set; }
+        public double Volume { get; set; }
         [JsonPropertyName("x")]
-        public int X { get; set; }
+        public double X { get; set; }
         [JsonPropertyName("y")]
-        public int Y { get; set; }
+        public double Y { get; set; }
         [JsonPropertyName("parentGroupName")]
         public string ParentGroupName { get; set; }
         [JsonPropertyName("groupChildren")]
